server: rafact the code structure
