+++ server/BUILD.bazel
@@ -3,7 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "server",
     srcs = [
-        "buffered_read_conn.go",
         "conn.go",
         "conn_stmt.go",
         "driver.go",
@@ -56,6 +55,7 @@ go_library(
         "//server/internal/column",
         "//server/internal/dump",
         "//server/internal/parse",
+        "//server/internal/parse/handshake",
         "//server/internal/util",
         "//server/metrics",
         "//session",
@@ -185,6 +185,8 @@ go_test(
         "//parser/mysql",
         "//planner/core",
         "//server/internal/column",
+        "//server/internal/parse",
+        "//server/internal/parse/handshake",
         "//server/internal/util",
         "//session",
         "//sessionctx",
