+++ server/conn_test.go
@@ -36,6 +36,9 @@ import (
 	"github.com/pingcap/tidb/extension"
 	"github.com/pingcap/tidb/parser/auth"
 	"github.com/pingcap/tidb/parser/mysql"
+	"github.com/pingcap/tidb/server/internal"
+	"github.com/pingcap/tidb/server/internal/parse"
+	"github.com/pingcap/tidb/server/internal/parse/handshake"
 	util2 "github.com/pingcap/tidb/server/internal/util"
 	"github.com/pingcap/tidb/session"
 	"github.com/pingcap/tidb/sessionctx/variable"
@@ -185,8 +188,8 @@ func TestIssue33699(t *testing.T) {
 
 func TestMalformHandshakeHeader(t *testing.T) {
 	data := []byte{0x00}
-	var p handshakeResponse41
-	_, err := parseHandshakeResponseHeader(context.Background(), &p, data)
+	var p handshake.HandshakeResponse41
+	_, err := parse.HandshakeResponseHeader(context.Background(), &p, data)
 	require.Error(t, err)
 }
 
@@ -206,11 +209,11 @@ func TestParseHandshakeResponse(t *testing.T) {
 		0x6c, 0x61, 0x74, 0x66, 0x6f, 0x72, 0x6d, 0x06, 0x78, 0x38, 0x36, 0x5f, 0x36, 0x34, 0x03, 0x66,
 		0x6f, 0x6f, 0x03, 0x62, 0x61, 0x72,
 	}
-	var p handshakeResponse41
-	offset, err := parseHandshakeResponseHeader(context.Background(), &p, data)
+	var p handshake.HandshakeResponse41
+	offset, err := parse.HandshakeResponseHeader(context.Background(), &p, data)
 	require.NoError(t, err)
 	require.Equal(t, mysql.ClientConnectAtts, p.Capability&mysql.ClientConnectAtts)
-	err = parseHandshakeResponseBody(context.Background(), &p, data, offset)
+	err = parse.HandshakeResponseBody(context.Background(), &p, data, offset)
 	require.NoError(t, err)
 	eq := mapIdentical(p.Attrs, map[string]string{
 		"_client_version": "5.6.6-m9",
@@ -229,15 +232,15 @@ func TestParseHandshakeResponse(t *testing.T) {
 		0x74, 0x00, 0x6d, 0x79, 0x73, 0x71, 0x6c, 0x5f, 0x6e, 0x61, 0x74, 0x69, 0x76, 0x65, 0x5f, 0x70,
 		0x61, 0x73, 0x73, 0x77, 0x6f, 0x72, 0x64, 0x00,
 	}
-	p = handshakeResponse41{}
-	offset, err = parseHandshakeResponseHeader(context.Background(), &p, data)
+	p = handshake.HandshakeResponse41{}
+	offset, err = parse.HandshakeResponseHeader(context.Background(), &p, data)
 	require.NoError(t, err)
 	capability := mysql.ClientProtocol41 |
 		mysql.ClientPluginAuth |
 		mysql.ClientSecureConnection |
 		mysql.ClientConnectWithDB
 	require.Equal(t, capability, p.Capability&capability)
-	err = parseHandshakeResponseBody(context.Background(), &p, data, offset)
+	err = parse.HandshakeResponseBody(context.Background(), &p, data, offset)
 	require.NoError(t, err)
 	require.Equal(t, "pam", p.User)
 	require.Equal(t, "test", p.DBName)
@@ -260,38 +263,13 @@ func TestIssue1768(t *testing.T) {
 		0x34, 0x0c, 0x70, 0x72, 0x6f, 0x67, 0x72, 0x61, 0x6d, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x05, 0x6d,
 		0x79, 0x73, 0x71, 0x6c,
 	}
-	p := handshakeResponse41{}
-	offset, err := parseHandshakeResponseHeader(context.Background(), &p, data)
+	p := handshake.HandshakeResponse41{}
+	offset, err := parse.HandshakeResponseHeader(context.Background(), &p, data)
 	require.NoError(t, err)
 	require.Equal(t, mysql.ClientPluginAuthLenencClientData, p.Capability&mysql.ClientPluginAuthLenencClientData)
-	err = parseHandshakeResponseBody(context.Background(), &p, data, offset)
+	err = parse.HandshakeResponseBody(context.Background(), &p, data, offset)
 	require.NoError(t, err)
 	require.NotEmpty(t, p.Auth)
-}
-
-func TestAuthSwitchRequest(t *testing.T) {
-	// this data is from a MySQL 8.0 client
-	data := []byte{
-		0x85, 0xa6, 0xff, 0x1, 0x0, 0x0, 0x0, 0x1, 0x21, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-		0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x72, 0x6f,
-		0x6f, 0x74, 0x0, 0x0, 0x63, 0x61, 0x63, 0x68, 0x69, 0x6e, 0x67, 0x5f, 0x73, 0x68, 0x61,
-		0x32, 0x5f, 0x70, 0x61, 0x73, 0x73, 0x77, 0x6f, 0x72, 0x64, 0x0, 0x79, 0x4, 0x5f, 0x70,
-		0x69, 0x64, 0x5, 0x37, 0x37, 0x30, 0x38, 0x36, 0x9, 0x5f, 0x70, 0x6c, 0x61, 0x74, 0x66,
-		0x6f, 0x72, 0x6d, 0x6, 0x78, 0x38, 0x36, 0x5f, 0x36, 0x34, 0x3, 0x5f, 0x6f, 0x73, 0x5,
-		0x4c, 0x69, 0x6e, 0x75, 0x78, 0xc, 0x5f, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x5f, 0x6e,
-		0x61, 0x6d, 0x65, 0x8, 0x6c, 0x69, 0x62, 0x6d, 0x79, 0x73, 0x71, 0x6c, 0x7, 0x6f, 0x73,
-		0x5f, 0x75, 0x73, 0x65, 0x72, 0xa, 0x6e, 0x75, 0x6c, 0x6c, 0x6e, 0x6f, 0x74, 0x6e, 0x69,
-		0x6c, 0xf, 0x5f, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x5f, 0x76, 0x65, 0x72, 0x73, 0x69,
-		0x6f, 0x6e, 0x6, 0x38, 0x2e, 0x30, 0x2e, 0x32, 0x31, 0xc, 0x70, 0x72, 0x6f, 0x67, 0x72,
-		0x61, 0x6d, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x5, 0x6d, 0x79, 0x73, 0x71, 0x6c,
-	}
-
-	var resp handshakeResponse41
-	pos, err := parseHandshakeResponseHeader(context.Background(), &resp, data)
-	require.NoError(t, err)
-	err = parseHandshakeResponseBody(context.Background(), &resp, data, pos)
-	require.NoError(t, err)
-	require.Equal(t, "caching_sha2_password", resp.AuthPlugin)
 }
 
 func TestInitialHandshake(t *testing.T) {
@@ -1177,7 +1155,7 @@ func TestHandleAuthPlugin(t *testing.T) {
 		server: srv,
 		user:   "unativepassword",
 	}
-	resp := handshakeResponse41{
+	resp := handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41 | mysql.ClientPluginAuth,
 		AuthPlugin: mysql.AuthNativePassword,
 	}
@@ -1198,7 +1176,7 @@ func TestHandleAuthPlugin(t *testing.T) {
 		server: srv,
 		user:   "unativepassword",
 	}
-	resp = handshakeResponse41{
+	resp = handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41 | mysql.ClientPluginAuth,
 		AuthPlugin: mysql.AuthCachingSha2Password,
 	}
@@ -1221,7 +1199,7 @@ func TestHandleAuthPlugin(t *testing.T) {
 		server: srv,
 		user:   "unativepassword",
 	}
-	resp = handshakeResponse41{
+	resp = handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41 | mysql.ClientPluginAuth,
 		AuthPlugin: mysql.AuthTiDBSM3Password,
 	}
@@ -1243,7 +1221,7 @@ func TestHandleAuthPlugin(t *testing.T) {
 		server: srv,
 		user:   "unativepassword",
 	}
-	resp = handshakeResponse41{
+	resp = handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41,
 	}
 	err = cc.handleAuthPlugin(ctx, &resp)
@@ -1266,7 +1244,7 @@ func TestHandleAuthPlugin(t *testing.T) {
 		server: srv,
 		user:   "unativepassword",
 	}
-	resp = handshakeResponse41{
+	resp = handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41 | mysql.ClientPluginAuth,
 		AuthPlugin: mysql.AuthNativePassword,
 	}
@@ -1289,7 +1267,7 @@ func TestHandleAuthPlugin(t *testing.T) {
 		server: srv,
 		user:   "unativepassword",
 	}
-	resp = handshakeResponse41{
+	resp = handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41 | mysql.ClientPluginAuth,
 		AuthPlugin: mysql.AuthCachingSha2Password,
 	}
@@ -1312,7 +1290,7 @@ func TestHandleAuthPlugin(t *testing.T) {
 		server: srv,
 		user:   "unativepassword",
 	}
-	resp = handshakeResponse41{
+	resp = handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41 | mysql.ClientPluginAuth,
 		AuthPlugin: mysql.AuthTiDBSM3Password,
 	}
@@ -1334,7 +1312,7 @@ func TestHandleAuthPlugin(t *testing.T) {
 		server: srv,
 		user:   "unativepassword",
 	}
-	resp = handshakeResponse41{
+	resp = handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41,
 	}
 	err = cc.handleAuthPlugin(ctx, &resp)
@@ -1358,7 +1336,7 @@ func TestHandleAuthPlugin(t *testing.T) {
 		server: srv,
 		user:   "unativepassword",
 	}
-	resp = handshakeResponse41{
+	resp = handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41 | mysql.ClientPluginAuth,
 		AuthPlugin: mysql.AuthNativePassword,
 	}
@@ -1381,7 +1359,7 @@ func TestHandleAuthPlugin(t *testing.T) {
 		server: srv,
 		user:   "unativepassword",
 	}
-	resp = handshakeResponse41{
+	resp = handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41 | mysql.ClientPluginAuth,
 		AuthPlugin: mysql.AuthCachingSha2Password,
 	}
@@ -1404,7 +1382,7 @@ func TestHandleAuthPlugin(t *testing.T) {
 		server: srv,
 		user:   "unativepassword",
 	}
-	resp = handshakeResponse41{
+	resp = handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41 | mysql.ClientPluginAuth,
 		AuthPlugin: mysql.AuthTiDBSM3Password,
 	}
@@ -1426,7 +1404,7 @@ func TestHandleAuthPlugin(t *testing.T) {
 		server: srv,
 		user:   "unativepassword",
 	}
-	resp = handshakeResponse41{
+	resp = handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41,
 	}
 	err = cc.handleAuthPlugin(ctx, &resp)
@@ -1450,7 +1428,7 @@ func TestHandleAuthPlugin(t *testing.T) {
 		server: srv,
 		user:   "unativepassword",
 	}
-	resp = handshakeResponse41{
+	resp = handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41 | mysql.ClientPluginAuth,
 		AuthPlugin: mysql.AuthNativePassword,
 	}
@@ -1473,7 +1451,7 @@ func TestHandleAuthPlugin(t *testing.T) {
 		server: srv,
 		user:   "unativepassword",
 	}
-	resp = handshakeResponse41{
+	resp = handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41 | mysql.ClientPluginAuth,
 		AuthPlugin: mysql.AuthCachingSha2Password,
 	}
@@ -1496,7 +1474,7 @@ func TestHandleAuthPlugin(t *testing.T) {
 		server: srv,
 		user:   "unativepassword",
 	}
-	resp = handshakeResponse41{
+	resp = handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41 | mysql.ClientPluginAuth,
 		AuthPlugin: mysql.AuthTiDBSM3Password,
 	}
@@ -1518,7 +1496,7 @@ func TestHandleAuthPlugin(t *testing.T) {
 		server: srv,
 		user:   "unativepassword",
 	}
-	resp = handshakeResponse41{
+	resp = handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41,
 	}
 	err = cc.handleAuthPlugin(ctx, &resp)
@@ -1605,7 +1583,7 @@ func TestAuthPlugin2(t *testing.T) {
 	}
 	cc.setCtx(tc)
 
-	resp := handshakeResponse41{
+	resp := handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41 | mysql.ClientPluginAuth,
 	}
 
@@ -1678,7 +1656,7 @@ func TestAuthSessionTokenPlugin(t *testing.T) {
 	tokenBytes := []byte(rows[0][1].(string))
 
 	// auth with the token
-	resp := handshakeResponse41{
+	resp := handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41 | mysql.ClientPluginAuth,
 		AuthPlugin: mysql.AuthTiDBSessionToken,
 		Auth:       tokenBytes,
@@ -1722,7 +1700,7 @@ func TestMaxAllowedPacket(t *testing.T) {
 	bytes := append([]byte{0x00, 0x04, 0x00, 0x00}, []byte(fmt.Sprintf("SELECT length('%s') as len;", strings.Repeat("a", 999)))...)
 	_, err := inBuffer.Write(bytes)
 	require.NoError(t, err)
-	brc := newBufferedReadConn(&bytesConn{inBuffer})
+	brc := internal.newBufferedReadConn(&bytesConn{inBuffer})
 	pkt := newPacketIO(brc)
 	pkt.setMaxAllowedPacket(maxAllowedPacket)
 	readBytes, err = pkt.readPacket()
@@ -1735,7 +1713,7 @@ func TestMaxAllowedPacket(t *testing.T) {
 	bytes = append([]byte{0x01, 0x04, 0x00, 0x00}, []byte(fmt.Sprintf("SELECT length('%s') as len;", strings.Repeat("a", 1000)))...)
 	_, err = inBuffer.Write(bytes)
 	require.NoError(t, err)
-	brc = newBufferedReadConn(&bytesConn{inBuffer})
+	brc = internal.newBufferedReadConn(&bytesConn{inBuffer})
 	pkt = newPacketIO(brc)
 	pkt.setMaxAllowedPacket(maxAllowedPacket)
 	_, err = pkt.readPacket()
@@ -1747,7 +1725,7 @@ func TestMaxAllowedPacket(t *testing.T) {
 	bytes = append([]byte{0x01, 0x02, 0x00, 0x00}, []byte(fmt.Sprintf("SELECT length('%s') as len;", strings.Repeat("a", 488)))...)
 	_, err = inBuffer.Write(bytes)
 	require.NoError(t, err)
-	brc = newBufferedReadConn(&bytesConn{inBuffer})
+	brc = internal.newBufferedReadConn(&bytesConn{inBuffer})
 	pkt = newPacketIO(brc)
 	pkt.setMaxAllowedPacket(maxAllowedPacket)
 	readBytes, err = pkt.readPacket()
@@ -1758,7 +1736,7 @@ func TestMaxAllowedPacket(t *testing.T) {
 	bytes = append([]byte{0x01, 0x02, 0x00, 0x01}, []byte(fmt.Sprintf("SELECT length('%s') as len;", strings.Repeat("b", 488)))...)
 	_, err = inBuffer.Write(bytes)
 	require.NoError(t, err)
-	brc = newBufferedReadConn(&bytesConn{inBuffer})
+	brc = internal.newBufferedReadConn(&bytesConn{inBuffer})
 	pkt.setBufferedReadConn(brc)
 	readBytes, err = pkt.readPacket()
 	require.NoError(t, err)
@@ -1970,7 +1948,7 @@ func TestAuthSha(t *testing.T) {
 	ctx := &TiDBContext{Session: tk.Session()}
 	cc.setCtx(ctx)
 
-	resp := handshakeResponse41{
+	resp := handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41 | mysql.ClientPluginAuth,
 		AuthPlugin: mysql.AuthCachingSha2Password,
 		Auth:       []byte{}, // No password
@@ -2047,7 +2025,7 @@ func TestLDAPAuthSwitch(t *testing.T) {
 	cc.isUnixSocket = true
 
 	require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/server/FakeAuthSwitch", "return(1)"))
-	respAuthSwitch, err := cc.checkAuthPlugin(context.Background(), &handshakeResponse41{
+	respAuthSwitch, err := cc.checkAuthPlugin(context.Background(), &handshake.HandshakeResponse41{
 		Capability: mysql.ClientProtocol41 | mysql.ClientPluginAuth,
 		User:       "test_simple_ldap",
 	})
