+++ server/internal/parse/BUILD.bazel
@@ -10,25 +10,33 @@ go_library(
         "//expression",
         "//parser/charset",
         "//parser/mysql",
+        "//server/internal/parse/handshake",
         "//server/internal/util",
         "//sessionctx/stmtctx",
         "//types",
         "//util/dbterror",
         "//util/hack",
+        "//util/logutil",
+        "@com_github_klauspost_compress//zstd",
+        "@org_uber_go_zap//:zap",
     ],
 )
 
 go_test(
     name = "parse_test",
     timeout = "short",
-    srcs = ["parse_test.go"],
+    srcs = [
+        "handshake_test.go",
+        "parse_test.go",
+    ],
     embed = [":parse"],
     flaky = True,
-    shard_count = 3,
+    shard_count = 4,
     deps = [
         "//expression",
         "//parser/mysql",
         "//parser/terror",
+        "//server/internal/parse/handshake",
         "//server/internal/util",
         "//sessionctx/stmtctx",
         "//types",
