+++ server/internal/util/buffered_read_conn.go
@@ -1,0 +1,39 @@
+// Copyright 2017 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package util
+
+import (
+	"bufio"
+	"net"
+)
+
+const DefaultReaderSize = 16 * 1024
+
+// BufferedReadConn is a net.Conn compatible structure that reads from bufio.Reader.
+type BufferedReadConn struct {
+	net.Conn
+	rb *bufio.Reader
+}
+
+func (conn BufferedReadConn) Read(b []byte) (n int, err error) {
+	return conn.rb.Read(b)
+}
+
+func NewBufferedReadConn(conn net.Conn) *BufferedReadConn {
+	return &BufferedReadConn{
+		Conn: conn,
+		rb:   bufio.NewReaderSize(conn, DefaultReaderSize),
+	}
+}
